Consolidate non-mermaid artifact tests in one file Fixes #11556
