+++ rust/kcl-lib/src/execution/artifact.rs
@@ -33,9 +33,9 @@ use crate::parsing::ast::types::Program;
 use crate::std::sketch::build_reverse_region_mapping;
 
 #[cfg(test)]
-mod entity_clone_tests;
-#[cfg(test)]
 mod mermaid_tests;
+#[cfg(test)]
+mod tests;
 
 macro_rules! internal_error {
     ($range:expr, $($rest:tt)*) => {{
