+++ rust/kcl-lib/src/execution/artifact/mermaid_tests.rs
@@ -1,4 +1,4 @@
-//! Tests for the artifact graph that convert it to Mermaid diagrams.
+//! Test-only helpers for rendering the artifact graph as Mermaid diagrams.
 use std::fmt::Write;
 
 use super::*;
@@ -661,336 +661,22 @@ impl ArtifactGraph {
 }
 
 #[test]
-fn surface_blend_creates_blend_sweep_artifact() {
-    let path_one_id = ArtifactId::new(Uuid::new_v4());
-    let path_two_id = ArtifactId::new(Uuid::new_v4());
-    let source_surface_one_id = ArtifactId::new(Uuid::new_v4());
-    let source_surface_two_id = ArtifactId::new(Uuid::new_v4());
-    let source_code_ref = CodeRef::placeholder(SourceRange::synthetic());
+fn pattern_traversal_links_source_and_copied_geometry() {
+    let source_id = ArtifactId::new(Uuid::new_v4());
+    let copy_id = ArtifactId::new(Uuid::new_v4());
+    let copy_face_id = ArtifactId::new(Uuid::new_v4());
+    let copy_edge_id = ArtifactId::new(Uuid::new_v4());
 
-    let mut artifacts = IndexMap::new();
-    artifacts.insert(
-        source_surface_one_id,
-        Artifact::Sweep(Sweep {
-            id: source_surface_one_id,
-            sub_type: SweepSubType::Extrusion,
-            path_id: path_one_id,
-            surface_ids: Vec::new(),
-            edge_ids: Vec::new(),
-            code_ref: source_code_ref.clone(),
-            trajectory_id: None,
-            method: kittycad_modeling_cmds::shared::ExtrudeMethod::Merge,
-            consumed: false,
-            pattern_ids: Vec::new(),
-        }),
-    );
-    artifacts.insert(
-        source_surface_two_id,
-        Artifact::Sweep(Sweep {
-            id: source_surface_two_id,
-            sub_type: SweepSubType::Extrusion,
-            path_id: path_two_id,
-            surface_ids: Vec::new(),
-            edge_ids: Vec::new(),
-            code_ref: source_code_ref,
-            trajectory_id: None,
-            method: kittycad_modeling_cmds::shared::ExtrudeMethod::Merge,
-            consumed: false,
-            pattern_ids: Vec::new(),
-        }),
-    );
+    let artifact = Artifact::Pattern(Pattern {
+        id: ArtifactId::new(Uuid::new_v4()),
+        sub_type: PatternSubType::Circular,
+        source_id,
+        copy_ids: vec![copy_id],
+        copy_face_ids: vec![copy_face_id],
+        copy_edge_ids: vec![copy_edge_id],
+        code_ref: CodeRef::placeholder(SourceRange::synthetic()),
+    });
 
-    let cmd_id = Uuid::new_v4();
-    let command = ModelingCmd::from(
-        kcmc::each_cmd::SurfaceBlend::builder()
-            .surfaces(vec![
-                kcmc::shared::SurfaceEdgeReference::builder()
-                    .object_id(Uuid::from(source_surface_one_id))
-                    .edges(vec![
-                        kcmc::shared::FractionOfEdge::builder()
-                            .edge_id(Uuid::new_v4())
-                            .lower_bound(0.0)
-                            .upper_bound(1.0)
-                            .build(),
-                    ])
-                    .build(),
-                kcmc::shared::SurfaceEdgeReference::builder()
-                    .object_id(Uuid::from(source_surface_two_id))
-                    .edges(vec![
-                        kcmc::shared::FractionOfEdge::builder()
-                            .edge_id(Uuid::new_v4())
-                            .lower_bound(0.0)
-                            .upper_bound(1.0)
-                            .build(),
-                    ])
-                    .build(),
-            ])
-            .build(),
-    );
-    let artifact_command = ArtifactCommand {
-        cmd_id,
-        range: SourceRange::synthetic(),
-        command,
-    };
-    let ast = crate::parsing::parse_str("", ModuleId::default()).unwrap();
-    let programs = crate::execution::ProgramLookup::new(ast, Default::default());
-
-    let updated = artifacts_to_update(
-        &artifacts,
-        &artifact_command,
-        &FnvHashMap::default(),
-        &FnvHashMap::default(),
-        &FnvHashMap::default(),
-        &programs,
-        0,
-        &IndexMap::default(),
-        &FnvHashMap::default(),
-    )
-    .unwrap();
-
-    assert_eq!(updated.len(), 1);
-    let Artifact::Sweep(blend_sweep) = &updated[0] else {
-        panic!("Expected SurfaceBlend to create a sweep artifact, got: {updated:?}");
-    };
-
-    assert_eq!(blend_sweep.id, ArtifactId::new(cmd_id));
-    assert_eq!(blend_sweep.sub_type, SweepSubType::Blend);
-    assert_eq!(blend_sweep.path_id, path_one_id);
-    assert_eq!(blend_sweep.trajectory_id, Some(path_two_id));
-    assert_eq!(blend_sweep.method, kittycad_modeling_cmds::shared::ExtrudeMethod::New);
-    assert!(!blend_sweep.consumed);
-}
-
-#[test]
-fn create_region_creates_region_path_sub_type() {
-    let origin_path_id = ArtifactId::new(Uuid::new_v4());
-    let origin_plane_id = ArtifactId::new(Uuid::new_v4());
-    let source_code_ref = CodeRef::placeholder(SourceRange::synthetic());
-
-    let mut artifacts = IndexMap::new();
-    artifacts.insert(
-        origin_path_id,
-        Artifact::Path(Path {
-            id: origin_path_id,
-            sub_type: PathSubType::Sketch,
-            plane_id: origin_plane_id,
-            seg_ids: Vec::new(),
-            consumed: false,
-            sweep_id: None,
-            trajectory_sweep_id: None,
-            solid2d_id: None,
-            code_ref: source_code_ref,
-            composite_solid_id: None,
-            sketch_block_id: None,
-            origin_path_id: None,
-            inner_path_id: None,
-            outer_path_id: None,
-            pattern_ids: Vec::new(),
-        }),
-    );
-
-    let cmd_id = Uuid::new_v4();
-    let command = ModelingCmd::from(
-        kcmc::each_cmd::CreateRegion::builder()
-            .object_id(Uuid::from(origin_path_id))
-            .segment(Uuid::new_v4())
-            .intersection_segment(Uuid::new_v4())
-            .intersection_index(-1)
-            .curve_clockwise(false)
-            .build(),
-    );
-    let artifact_command = ArtifactCommand {
-        cmd_id,
-        range: SourceRange::synthetic(),
-        command,
-    };
-    let ast = crate::parsing::parse_str("", ModuleId::default()).unwrap();
-    let programs = crate::execution::ProgramLookup::new(ast, Default::default());
-
-    let updated = artifacts_to_update(
-        &artifacts,
-        &artifact_command,
-        &FnvHashMap::default(),
-        &FnvHashMap::default(),
-        &FnvHashMap::default(),
-        &programs,
-        0,
-        &IndexMap::default(),
-        &FnvHashMap::default(),
-    )
-    .unwrap();
-
-    assert_eq!(updated.len(), 1);
-    let Artifact::Path(region_path) = &updated[0] else {
-        panic!("Expected CreateRegion to create a path artifact, got: {updated:?}");
-    };
-    assert_eq!(region_path.id, ArtifactId::new(cmd_id));
-    assert_eq!(region_path.sub_type, PathSubType::Region);
-    assert_eq!(region_path.plane_id, origin_plane_id);
-    // A region path isn't created from a sketch block directly.
-    assert_eq!(region_path.sketch_block_id, None);
-    // It links back to the origin sketch path.
-    assert_eq!(region_path.origin_path_id, Some(origin_path_id));
-}
-
-#[test]
-fn pattern_artifact_links_to_source_geometry() {
-    let path_id = ArtifactId::new(Uuid::new_v4());
-    let sweep_id = ArtifactId::new(Uuid::new_v4());
-    let pattern_id = ArtifactId::new(Uuid::new_v4());
-    let plane_id = ArtifactId::new(Uuid::new_v4());
-    let copy_id = Uuid::new_v4();
-    let copy_face_id = Uuid::new_v4();
-    let copy_edge_id = Uuid::new_v4();
-    let code_ref = CodeRef::placeholder(SourceRange::synthetic());
-    let face_edge_infos: Vec<kcmc::output::FaceEdgeInfo> = serde_json::from_value(serde_json::json!([
-        {
-            "object_id": copy_id,
-            "faces": [copy_face_id],
-            "edges": [copy_edge_id],
-        }
-    ]))
-    .expect("valid face-edge info");
-
-    let mut artifacts = IndexMap::new();
-    artifacts.insert(
-        path_id,
-        Artifact::Path(Path {
-            id: path_id,
-            sub_type: PathSubType::Sketch,
-            plane_id,
-            seg_ids: Vec::new(),
-            consumed: true,
-            sweep_id: Some(sweep_id),
-            trajectory_sweep_id: None,
-            solid2d_id: None,
-            code_ref: code_ref.clone(),
-            composite_solid_id: None,
-            sketch_block_id: None,
-            origin_path_id: None,
-            inner_path_id: None,
-            outer_path_id: None,
-            pattern_ids: Vec::new(),
-        }),
-    );
-    artifacts.insert(
-        sweep_id,
-        Artifact::Sweep(Sweep {
-            id: sweep_id,
-            sub_type: SweepSubType::Extrusion,
-            path_id,
-            surface_ids: Vec::new(),
-            edge_ids: Vec::new(),
-            code_ref: code_ref.clone(),
-            trajectory_id: None,
-            method: kittycad_modeling_cmds::shared::ExtrudeMethod::Merge,
-            consumed: false,
-            pattern_ids: Vec::new(),
-        }),
-    );
-
-    let updated = pattern_artifact_updates(
-        &artifacts,
-        pattern_id,
-        PatternSubType::Circular,
-        path_id,
-        &face_edge_infos,
-        code_ref,
-    );
-
-    assert!(matches!(
-        updated.first(),
-        Some(Artifact::Pattern(Pattern {
-            id,
-            sub_type: PatternSubType::Circular,
-            source_id,
-            copy_ids,
-            copy_face_ids,
-            copy_edge_ids,
-            ..
-        })) if *id == pattern_id
-            && *source_id == path_id
-            && copy_ids == &vec![ArtifactId::new(copy_id)]
-            && copy_face_ids == &vec![ArtifactId::new(copy_face_id)]
-            && copy_edge_ids == &vec![ArtifactId::new(copy_edge_id)]
-    ));
-    assert_eq!(
-        updated.first().map(Artifact::child_ids),
-        Some(vec![
-            ArtifactId::new(copy_id),
-            ArtifactId::new(copy_face_id),
-            ArtifactId::new(copy_edge_id)
-        ])
-    );
-    assert!(updated.iter().any(|artifact| {
-        matches!(artifact, Artifact::Path(path) if path.id == path_id && path.pattern_ids == vec![pattern_id])
-    }));
-    assert!(updated.iter().any(|artifact| {
-        matches!(artifact, Artifact::Sweep(sweep) if sweep.id == sweep_id && sweep.pattern_ids == vec![pattern_id])
-    }));
-}
-
-#[test]
-fn primitive_edge_does_not_replace_existing_segment_artifact() {
-    let shared_id = ArtifactId::new(Uuid::new_v4());
-    let path_id = ArtifactId::new(Uuid::new_v4());
-    let solid_id = ArtifactId::new(Uuid::new_v4());
-
-    let mut map = IndexMap::new();
-    map.insert(
-        shared_id,
-        Artifact::Segment(Segment {
-            id: shared_id,
-            path_id,
-            original_seg_id: None,
-            surface_id: None,
-            edge_ids: Vec::new(),
-            edge_cut_id: None,
-            code_ref: CodeRef::placeholder(SourceRange::synthetic()),
-            common_surface_ids: Vec::new(),
-        }),
-    );
-
-    merge_artifact_into_map(
-        &mut map,
-        Artifact::PrimitiveEdge(PrimitiveEdge {
-            id: shared_id,
-            solid_id,
-            code_ref: CodeRef::placeholder(SourceRange::synthetic()),
-        }),
-    );
-
-    assert!(matches!(map.get(&shared_id), Some(Artifact::Segment(_))));
-}
-
-#[test]
-fn primitive_face_does_not_replace_existing_cap_artifact() {
-    let shared_id = ArtifactId::new(Uuid::new_v4());
-    let sweep_id = ArtifactId::new(Uuid::new_v4());
-    let solid_id = ArtifactId::new(Uuid::new_v4());
-
-    let mut map = IndexMap::new();
-    map.insert(
-        shared_id,
-        Artifact::Cap(Cap {
-            id: shared_id,
-            sub_type: CapSubType::End,
-            edge_cut_edge_ids: Vec::new(),
-            sweep_id,
-            path_ids: Vec::new(),
-            face_code_ref: CodeRef::placeholder(SourceRange::synthetic()),
-            cmd_id: Uuid::new_v4(),
-        }),
-    );
-
-    merge_artifact_into_map(
-        &mut map,
-        Artifact::PrimitiveFace(PrimitiveFace {
-            id: shared_id,
-            solid_id,
-            code_ref: CodeRef::placeholder(SourceRange::synthetic()),
-        }),
-    );
-
-    assert!(matches!(map.get(&shared_id), Some(Artifact::Cap(_))));
+    assert_eq!(artifact.back_edges(), vec![source_id]);
+    assert_eq!(artifact.child_ids(), vec![copy_id, copy_face_id, copy_edge_id]);
 }
